requirements.txt
streamlit==1.28.0
Pillow==10.0.0
requests==2.31.0
plotly>=5.0.0
streamlit-lottie>=0.0.5
